Support MC flag as used by biobambam
biobambam (https://github.com/gt1/biobambam) uses the MC SAM flag to
store the coordinate of the mate in their duplicate marking. If using
biobambam after samblaster (for merging/sorting/de-duplicating split
BAMs) this overwrites samblaster's flag. This commit avoids failing
in this case since the original code assumed a string, and makes use
of this if present.