Change parameters for duplication genotyping model
This relaxes the allele balance expectation for duplications.  For a
simple tandem duplication, we expect to see 1/5 of the reads supporting
the duplication in the heterozygous state and 1/3 supporting the
duplication in the homozygous state. Additional copies of the duplicon
will alter this ratio, but this is not handled by the previous or
current model.